bugfixes for rendering creators
