+++ frontend/src/pages/CreatorsPage.jsx
@@ -1,7 +1,9 @@
 import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 
 import CreatorsList from '../components/creators/CreatorsList';
+import DefaultButton from '../components/misc/DefaultButton';
 import ImageTitle from '../components/misc/ImageTitle';
 import creatorsService from '../services/creators.service';
 
@@ -9,6 +11,7 @@ export default function CreatorsPage() {
   const [creators, setCreators] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
+  const navigate = useNavigate();
 
   const fetchCreators = async () => {
     try {
@@ -27,6 +30,14 @@ export default function CreatorsPage() {
   useEffect(() => {
     fetchCreators();
   }, []);
+
+  const handleBack = () => {
+    if (window.history.length > 2) {
+      navigate(-1);
+    } else {
+      navigate('/');
+    }
+  };
 
   if (loading) {
     return (
@@ -55,6 +66,9 @@ export default function CreatorsPage() {
       <ImageTitle title="Alkotók" description="Alkotók, akik a BreakThe4th partnerei!" />
       <div className="min-h-screen flex flex-col items-center justify-center ">
         <CreatorsList creators={creators} />
+      </div>
+      <div className="mt-10 flex justify-center mb-10">
+        <DefaultButton onClick={handleBack} text="Vissza" />
       </div>
     </>
   );
